I started integrate new logic for single node

+++ worker/src/main/java/com/rakovpublic/jneuropallium/worker/neuron/INeuron.java
@@ -156,7 +156,7 @@ public interface INeuron extends Serializable, Comparable {
 
      void setCyclingNeuronInputMapping(HashMap<String, Long> cyclingNeuronInputMapping);
 
-     Integer getCurrentLoopAmount();
+     Long getCurrentLoopAmount();
 
-     void setCurrentLoopAmount(Integer currentLoopCount) ;
+     void setCurrentLoopAmount(Long currentLoopCount) ;
 }
